refactor(zsh): simplify conditional checks

+++ plugins/golang/golang.plugin.zsh
@@ -17,9 +17,7 @@ if exists go; then
 
     command go mod init "$namespace"
 
-    if [[ ! -f main.go ]]; then
-      command cp -- "$ZDOTFILES_DIR/plugins/golang/main.go" .
-    fi
+    [[ -f main.go ]] || command cp -- "$ZDOTFILES_DIR/plugins/golang/main.go" .
   }
 
   if exists brew; then
